+++ src/main/java/kitchenpos/application/dto/response/MenuProductResponse.java
@@ -2,7 +2,7 @@ package kitchenpos.application.dto.response;
 
 import java.util.List;
 import java.util.stream.Collectors;
-import kitchenpos.domain.MenuProduct;
+import kitchenpos.menu.domain.MenuProduct;
 
 public class MenuProductResponse {
