+++ src/main/java/kitchenpos/menu/domain/Menu.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.menu.domain;
 
 import java.math.BigDecimal;
 import java.util.ArrayList;
@@ -15,6 +15,7 @@ import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 import javax.persistence.OneToMany;
 import kitchenpos.common.vo.Price;
+import kitchenpos.domain.MenuGroup;
 
 @Entity
 public class Menu {
