+++ src/main/java/kitchenpos/menu/domain/MenuProduct.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.menu.domain;
 
 import static java.util.Objects.nonNull;
