+++ src/main/java/kitchenpos/menu/domain/MenuProductRepository.java
@@ -1,6 +1,5 @@
-package kitchenpos.repository;
+package kitchenpos.menu.domain;
 
-import kitchenpos.domain.MenuProduct;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
