+++ src/test/java/kitchenpos/application/MenuServiceTest.java
@@ -12,6 +12,7 @@ import kitchenpos.application.dto.response.MenuGroupResponse;
 import kitchenpos.application.dto.response.MenuResponse;
 import kitchenpos.application.dto.response.ProductResponse;
 import kitchenpos.common.vo.Price;
+import kitchenpos.menu.service.MenuService;
 import kitchenpos.product.service.ProductService;
 import kitchenpos.supports.IntegrationTest;
 import kitchenpos.supports.MenuFixture;
