+++ src/test/java/kitchenpos/application/TableGroupServiceTest.java
@@ -13,6 +13,7 @@ import kitchenpos.application.dto.response.OrderResponse;
 import kitchenpos.application.dto.response.ProductResponse;
 import kitchenpos.application.dto.response.TableGroupResponse;
 import kitchenpos.application.dto.response.TableResponse;
+import kitchenpos.menu.service.MenuService;
 import kitchenpos.product.service.ProductService;
 import kitchenpos.supports.IntegrationTest;
 import kitchenpos.supports.MenuFixture;
